refactor: 각 액티비티를 패키지로 구성
