+++ app/src/main/java/com/moroom/android/ui/main/MainActivity.java
@@ -1,4 +1,4 @@
-package com.moroom.android.ui;
+package com.moroom.android.ui.main;
 
 import android.os.Bundle;
 import android.widget.Toast;
