+++ app/src/main/java/com/moroom/android/ui/navui/home/HomeFragment.kt
@@ -17,9 +17,9 @@ import com.google.firebase.auth.FirebaseAuth
 
 import com.moroom.android.ui.login.MoveToLogin
 import com.moroom.android.R
-import com.moroom.android.ui.ResultActivity
-import com.moroom.android.ui.SearchActivity
-import com.moroom.android.ui.WriteActivity
+import com.moroom.android.ui.result.ResultActivity
+import com.moroom.android.ui.search.SearchActivity
+import com.moroom.android.ui.write.WriteActivity
 import com.moroom.android.databinding.FragmentHomeBinding
 
 class HomeFragment : Fragment() {
