+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".ui.SignupActivity">
+    tools:context=".ui.signup.SignupActivity">
 
     <EditText
         android:id="@+id/et_emailAddress"
